refactor: 카드머신 도메인을 제거함
