+++ README.md
@@ -33,12 +33,13 @@
 - [x] 카드
   - [x] 문양을 가진다.
   - [x] 값을 가진다.
-  - [ ] 자신의 문양을 반환한다.
+  - [x] 자신의 문양을 반환한다.
   - [x] 자신의 값을 반환한다.
 
-- [x] 카드들
-  - [x] 중복되지 않는 카드들을 가진다. --> 어떻게 테스트 하지?
-  - [x] 카드를 한번 섞는다. --> 어떻게 테스트 하지?
+- [x] 카드덱
+  - [x] 중복되지 않는 카드들을 가진다. --> 어떻게 테스트 하지? --> 테스트를 해야하나?
+  - [x] 카드를 한번 섞는다. --> 어떻게 테스트 하지? --> 테스트를 해야하나?
+  - [x] 게임 시작 시 게임 참여자에게 카드를 두 장씩 배분한다.
   - [x] 어느 카드까지 뽑았는지 저장한다.
   - [x] 현재 인덱스가 가르키는 카드를 뽑아서 반환한다.
 
@@ -60,9 +61,7 @@
   - [x] 참여자들은 1명이상 10명 이하다
   - [x] 참여자들의 이름은 중복될 수 없다.
 
-- [x] 카드 머신
-  - [x] 게임 시작 시 게임 참여자에게 카드를 두 장씩 배분한다.
-  - [x] 한 명의 참여자에게 한 장의 카드를 배분한다.
+- ~~[x] 카드 머신~~
   
 - [x] 게임 결과
   - [x] 게스트현의 점수가 21을 초과할 경우 게스트가 패배한다.
