+++ src/main/java/blackjackgame/controller/blackJackController.java
@@ -3,7 +3,6 @@ package blackjackgame.controller;
 import java.util.List;
 
 import blackjackgame.domain.Result;
-import blackjackgame.domain.card.CardMachine;
 import blackjackgame.domain.card.Deck;
 import blackjackgame.domain.player.Dealer;
 import blackjackgame.domain.player.Guest;
@@ -24,11 +23,11 @@ public class blackJackController {
     public void run() {
         final Guests guests = generateGuests();
         final Dealer dealer = new Dealer();
-        final CardMachine cardMachine = new CardMachine(new Deck());
-        setGame(guests, dealer, cardMachine);
+        final Deck deck = new Deck();
+        setGame(guests, dealer, deck);
 
-        askGuestsHitCard(guests.getGuests(), cardMachine);
-        askDealerHitCard(dealer, cardMachine);
+        askGuestsHitCard(guests.getGuests(), deck);
+        askDealerHitCard(dealer, deck);
 
         printPlayersCardScore(guests, dealer);
         printGameResult(guests, dealer);
@@ -44,39 +43,39 @@ public class blackJackController {
         }
     }
 
-    private void setGame(final Guests guests, final Dealer dealer, final CardMachine cardMachine) {
-        cardMachine.initPlayersCards(guests, dealer);
+    private void setGame(final Guests guests, final Dealer dealer, final Deck deck) {
+        deck.initPlayersCards(guests, dealer);
         outputView.printFirstDealerCards(dealer.getName(), dealer.getCards());
         for (final Guest guest : guests.getGuests()) {
             outputView.printCards(guest.getName(), guest.getCards());
         }
     }
 
-    private void askGuestsHitCard(final List<Guest> guests, final CardMachine cardMachine) {
+    private void askGuestsHitCard(final List<Guest> guests, final Deck deck) {
         for (Guest guest : guests) {
-            askGuestHitCardRepeat(cardMachine, guest);
+            askGuestHitCardRepeat(deck, guest);
         }
     }
 
-    private void askGuestHitCardRepeat(final CardMachine cardMachine, final Guest guest) {
+    private void askGuestHitCardRepeat(final Deck deck, final Guest guest) {
         AddCardResponse addCardResponse = AddCardResponse.YES;
         while (guest.canHit() && addCardResponse == AddCardResponse.YES) {
             addCardResponse = inputView.readWantMoreCard(guest.getName());
-            askGuestHitCard(cardMachine, guest, addCardResponse);
+            askGuestHitCard(deck, guest, addCardResponse);
         }
     }
 
-    private void askGuestHitCard(final CardMachine cardMachine, final Guest guest,
+    private void askGuestHitCard(final Deck deck, final Guest guest,
         final AddCardResponse addCardResponse) {
         if (addCardResponse == AddCardResponse.YES) {
-            cardMachine.distributeCard(guest);
+            deck.distributeCard(guest);
         }
         outputView.printCards(guest.getName(), guest.getCards());
     }
 
-    private void askDealerHitCard(final Dealer dealer, final CardMachine cardMachine) {
+    private void askDealerHitCard(final Dealer dealer, final Deck deck) {
         while (dealer.canHit()) {
-            cardMachine.distributeCard(dealer);
+            deck.distributeCard(dealer);
             outputView.dealerAddCard();
         }
     }
